pyproject.toml
[tool.coverage.run]
omit = [
  "openshift_day2_configuration/cli.py",
  "openshift_day2_configuration/tests/*",
]

[tool.coverage.report]
fail_under = 47   # TODO: Need to be at least 60
skip_empty = true

[tool.coverage.html]
directory = ".tests_coverage"

[tool.ruff]
preview = true
line-length = 120
fix = true
output-format = "grouped"

[tool.ruff.format]
exclude = [".git", ".venv", ".mypy_cache", ".tox", "__pycache__"]

[tool.poetry]
name = "openshift-day2-configuration"
version = "0.1.0"
description = "Tool to configure specific day 2 operations on Openshift"
authors = ["myakove <myakove@gmail.com>", "rnetser <ruty_n4@gmail.com>"]
license = "Apache-2.0"
readme = "README.md"
packages = [{ include = "openshift_day2_configuration" }]
repository = "https://github.com/RedHatQE/openshift-day2-configuration"
homepage = "https://github.com/RedHatQE/openshift-day2-configuration"
documentation = "https://github.com/RedHatQE/openshift-day2-configuration/blob/main/README.md"
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
]

[tool.poetry.scripts]
openshift-cli-installer = "openshift_day2_configuration:main"

[tool.poetry.dependencies]
python = "^3.8"
pyaml-env = "^1.2.1"
python-simple-logger = "^1.0.19"
rich = "^13.7.1"
click = "^8.1.7"
openshift-python-wrapper = "^10.0.30"
openshift-python-utilities = "^5.0.35"
pyhelper-utils = "^0.0.7"


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
ipython = "*"


[tool.poetry.group.tests.dependencies]
pytest-cov = "^5.0.0"
pytest-mock = "^3.14.0"
pytest = "^8.1.1"

[tool.poetry-dynamic-versioning]
enable = true
pattern = "((?P<epoch>\\d+)!)?(?P<base>\\d+(\\.\\d+)*)"

[tool.poetry-dynamic-versioning.substitution]
files = ["VERSION"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
